requirements.txt
langchain
langchain-community
langchain-core
langchain_qdrant
langchain-text-splitters
protobuf
pydantic
pydantic_core
langchain_openai
qdrant-client
streamlit
psutil
pymupdf
python-dotenv